+++ docs/src/examples/components/Toolbar/Usage/ToolbarExampleActionPopupInMenu.shorthand.tsx
@@ -30,6 +30,7 @@ const ToolbarExampleActionPopupInMenu = () => {
   const [menuOpen, setMenuOpen] = React.useState(false)
   return (
     <Toolbar
+      aria-label="Popup with an action in menu"
       items={[
         {
           icon: 'more',