+++ docs/src/examples/components/Toolbar/Usage/ToolbarExampleWithTooltip.shorthand.tsx
@@ -19,6 +19,7 @@ const ToolbarExampleShorthand = () => {
 
   return (
     <Toolbar
+      aria-label="With tooltips"
       items={[
         {
           key: 'bold',