+++ docs/src/routes.tsx
@@ -45,6 +45,7 @@ import VirtualizedTreePrototype from './prototypes/VirtualizedTree'
 import CopyToClipboardPrototype from './prototypes/CopyToClipboard'
 import ParticipantsListPrototype from './prototypes/ParticipantsList'
 import CustomScrollbarPrototype from './prototypes/customScrollbar'
+import EditorToolbarPrototype from './prototypes/EditorToolbar'
 
 const Routes = () => (
   <BrowserRouter basename={__BASENAME__}>
@@ -75,6 +76,7 @@ const Routes = () => (
           <Route exact path="/prototype-popups" component={PopupsPrototype} />
           <Route exact path="/icon-viewer" component={IconViewerPrototype} />
           <Route exact path="/prototype-alerts" component={AlertsPrototype} />
+          <Route exact path="/prototype-editor-toolbar" component={EditorToolbarPrototype} />
           <Route
             exact
             path="/prototype-nested-popups-and-dialogs"