+++ packages/react/src/components/Toolbar/ToolbarItem.tsx
@@ -130,7 +130,10 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
     active: PropTypes.bool,
     disabled: PropTypes.bool,
     icon: customPropTypes.itemShorthandWithoutJSX,
-    menu: PropTypes.oneOfType([customPropTypes.itemShorthand, customPropTypes.collectionShorthand]),
+    menu: PropTypes.oneOfType([
+      customPropTypes.shorthandAllowingChildren,
+      PropTypes.arrayOf(customPropTypes.shorthandAllowingChildren),
+    ]),
     menuOpen: PropTypes.bool,
     onMenuOpenChange: PropTypes.func,
     onClick: PropTypes.func,
@@ -144,7 +147,7 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
       }),
       PropTypes.string,
     ]),
-    wrapper: customPropTypes.itemShorthand,
+    wrapper: customPropTypes.shorthandAllowingChildren,
   }
 
   static defaultProps = {
@@ -158,11 +161,12 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
       event.preventDefault()
       this.handleClick(event)
     },
+    performWrapperClick: event => {
+      this.handleWrapperClick(event)
+    },
     closeMenuAndFocusTrigger: event => {
       this.trySetMenuOpen(false, event)
-      if (this.itemRef) {
-        this.itemRef.current.focus()
-      }
+      _.invoke(this.itemRef.current, 'focus')
     },
     doNotNavigateNextToolbarItem: event => {
       event.stopPropagation()
@@ -170,8 +174,9 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
   }
 
   itemRef = React.createRef<HTMLElement>()
+  menuRef = React.createRef<HTMLElement>() as React.MutableRefObject<HTMLElement>
 
-  handleMenuOverrides = (getRefs: GetRefs, variables) => (predefinedProps: ToolbarItemProps) => ({
+  handleMenuOverrides = (getRefs: GetRefs, variables) => (predefinedProps: ToolbarMenuProps) => ({
     onBlur: (e: React.FocusEvent) => {
       const isInside = _.some(getRefs(), (childRef: NodeRef) => {
         return childRef.current.contains(e.relatedTarget as HTMLElement)
@@ -181,15 +186,16 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
         this.trySetMenuOpen(false, e)
       }
     },
-    onItemClick: (e, itemProps: ToolbarItemProps) => {
+    onItemClick: (e, itemProps: ToolbarMenuItemProps) => {
+      const { popup, menuOpen } = itemProps
       _.invoke(predefinedProps, 'onItemClick', e, itemProps)
-      if (itemProps.popup) {
+      if (popup) {
         return
       }
       // TODO: should we pass toolbarMenuItem to the user callback so he can decide if he wants to close the menu?
-      this.trySetMenuOpen(false, e)
-      if (this.itemRef) {
-        this.itemRef.current.focus()
+      this.trySetMenuOpen(menuOpen, e)
+      if (!menuOpen) {
+        _.invoke(this.itemRef.current, 'focus')
       }
     },
     variables: mergeComponentVariables(variables, predefinedProps.variables),
@@ -217,8 +223,22 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
       <Unstable_NestingAuto>
         {(getRefs, nestingRef) => (
           <>
-            <Ref innerRef={nestingRef}>
-              <Popper align="start" position="above" targetRef={this.itemRef}>
+            <Ref
+              innerRef={(node: HTMLElement) => {
+                nestingRef.current = node
+                this.menuRef.current = node
+              }}
+            >
+              <Popper
+                align="start"
+                position="above"
+                modifiers={{
+                  preventOverflow: {
+                    escapeWithReference: false, // escapeWithReference breaks positioning of ToolbarMenu in overflow mode because Popper components sets modifiers on scrollable container
+                  },
+                }}
+                targetRef={this.itemRef}
+              >
                 {ToolbarMenu.create(menu, {
                   overrideProps: this.handleMenuOverrides(getRefs, variables),
                 })}
@@ -263,8 +283,12 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
             ...accessibility.attributes.wrapper,
             ...applyAccessibilityKeyHandlers(accessibility.keyHandlers.wrapper, wrapper),
           },
-          overrideProps: () => ({
+          overrideProps: predefinedProps => ({
             children: contentElement,
+            onClick: e => {
+              this.handleWrapperClick(e)
+              _.invoke(predefinedProps, 'onClick', e)
+            },
           }),
         })
       }
@@ -296,6 +320,16 @@ class ToolbarItem extends UIComponent<WithAsProp<ToolbarItemProps>> {
     }
 
     _.invoke(this.props, 'onClick', e, this.props)
+  }
+
+  handleWrapperClick = e => {
+    const { menu } = this.props
+    if (menu) {
+      if (doesNodeContainClick(this.menuRef.current, e, this.context.target)) {
+        this.trySetMenuOpen(false, e)
+        _.invoke(this.itemRef.current, 'focus')
+      }
+    }
   }
 
   handleOutsideClick = (getRefs: GetRefs) => (e: MouseEvent) => {