+++ packages/react/src/lib/createComponent.ts
@@ -42,7 +42,7 @@ const createComponent = <P extends ObjectOf<any> = any>({
   const StardustComponent: CreateComponentReturnType<P> = (props): React.ReactElement<P> => {
     // Stores debug information for component.
     // Note that this ref should go as the first one, to be discoverable by debug utils.
-    const stardustDebug = React.useRef(null)
+    const ref = React.useRef(null)
 
     const context: ProviderContextPrepared = React.useContext(ThemeContext)
 
@@ -55,7 +55,7 @@ const createComponent = <P extends ObjectOf<any> = any>({
         state: {},
         actionHandlers,
         render: config => render(config, props),
-        saveDebug: updatedDebug => (stardustDebug.current = updatedDebug),
+        saveDebug: stardustDebug => (ref.current = { stardustDebug }),
       },
       context,
     )