+++ packages/react/src/themes/teams/components/Icon/svg/icons/index.ts
@@ -118,6 +118,7 @@ import participantRemove from './participantRemove'
 import phoneClock from './phoneClock'
 import phoneArrow from './phoneArrow'
 import play from './play'
+import plugs from './plugs'
 import poll from './poll'
 import presenceAvailable from './presenceAvailable'
 import presenceStroke from './presenceStroke'
@@ -130,6 +131,7 @@ import redo from './redo'
 import removeFormat from './removeFormat'
 import reply from './reply'
 import retry from './retry'
+import robot from './robot'
 import shareAlt from './shareAlt'
 import skypeLogo from './skypeLogo'
 import snooze from './snooze'
@@ -142,6 +144,7 @@ import stardustCircle from './stardustCircle'
 import sticker from './sticker'
 import strike from './strike'
 import sync from './sync'
+import tabs from './tabs'
 import table from './table'
 import tableAdd from './tableAdd'
 import tableDelete from './tableDelete'
@@ -276,6 +279,7 @@ export default {
   'phone-clock': phoneClock,
   'phone-arrow': phoneArrow,
   play,
+  plugs,
   poll,
   'presence-available': presenceAvailable,
   'presence-stroke': presenceStroke,
@@ -288,6 +292,7 @@ export default {
   'remove-format': removeFormat,
   reply,
   retry,
+  robot,
   search,
   send,
   settings,
@@ -296,6 +301,7 @@ export default {
   sticker,
   strike,
   sync,
+  tabs,
   table,
   'table-add': tableAdd,
   'table-delete': tableDelete,